requirements.txt
# torch==1.3.0
# torchvision==0.4.1
absl-py==0.10.0
alabaster==0.7.12
allennlp==0.9.0
anyio==2.0.2
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
async-generator==1.10
attrs==20.2.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bleach==1.5.0
blis==0.2.4
boto3==1.15.3
botocore==1.18.3
catalogue==2.0.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
conllu==1.3.1
contextvars==2.4
cycler==0.10.0
cymem==2.0.5
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
editdistance==0.5.3
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flaky==3.7.0
Flask==1.1.2
Flask-Cors==3.0.10
ftfy==5.8
gast==0.4.0
gevent==21.1.2
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
imagesize==1.2.0
immutables==0.14
importlib-metadata==2.0.0
iniconfig==1.0.1
ipdb==0.13.4
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonlines==1.2.0
jsonnet==0.17.0
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.2.3
jupyterlab==3.0.7
# jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.5
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kmeans-pytorch==0.3
# lang2vec==1.1.6
Markdown==3.2.2
MarkupSafe==1.1.1
# matplotlib==3.0.3
metrics==0.3.3
mistune==0.8.4
murmurhash==1.0.5
namedlist==1.8
nbclassic==0.2.6
nbclient==0.5.1
# nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
notebook==6.2.0
numpy==1.16.2
numpydoc==1.1.0
openpyxl==3.0.5
overrides==3.1.0
oyaml==1.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.3
parsimonious==0.8.1
parso==0.8.1
pathlib2==2.3.5
pathspec==0.5.5
pathy==0.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.13.0
ptyprocess==0.7.0
py==1.9.0
pyaml==20.4.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
Pygments==2.2.0
pynvml==8.0.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.0
python-dateutil==2.8.1
python-louvain==0.13
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==22.0.2
regex==2020.7.14
requests==2.24.0
responses==0.12.1
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scikit-optimize==0.8.1
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.91
seqeval==1.2.2
six==1.15.0
smart-open==3.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
spacy==2.1.9
spacy-legacy==3.0.1
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
srsly==1.0.5
tabulate==0.8.7
tensorboard==1.9.0
tensorboardX==1.9
tensorflow==1.9.0
tensorflow-hub==0.11.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
thinc==7.0.8
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
toml==0.10.1
torchsummary==1.5.1
tornado==6.1
tqdm==4.49.0
traitlets==4.3.3
transformers==3.2.0
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.25.10
utils==1.0.1
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
word2number==1.1
zipp==3.2.0
zope.event==4.5.0
zope.interface==5.2.0

